clean up comments on v2
